# Fix the unfair betwenn C and Rust code (unsigned 32 bit interger) and use stable toolchain