--- a/fibonacci/rust/code.rs
+++ b/fibonacci/rust/code.rs
@@ -1,5 +1,13 @@
+fn fibonacci(n: u32) -> u32 {
+    match n {
+        0 => 0,
+        1 => 1,
+        _ => fibonacci(n - 1) + fibonacci(n - 2),
+    }
+}
+
 fn main() {
-    let u: i32 = std::env::args()
+    let u: u32 = std::env::args()
         .nth(1)
         .unwrap()
         .parse()
@@ -9,12 +17,4 @@ fn main() {
         r += fibonacci(i);
     }
     println!("{}", r);
-}
-
-fn fibonacci(n: i32) -> i32 {
-    match n {
-        0 => 0,
-        1 => 1,
-        _ => fibonacci(n - 1) + fibonacci(n - 2),
-    }
 }