Hello,
I have fixed the rust performance problem, the original code was using size  as the size of integer for the tab A and the C code is using integer of 32 bits.
The size type is the type that matches the size of a pointer to refer to any location in memory, so 32 bits on 32 bit platform and 64 bits on most computer today.

I have fixed  that by just replacing usize by u32
So compiler llvm for both can optimize the code the same way.

On my machine
Before: 
C: 1.83 sec
Rust: 2.01 sec
After: 
C: 1.83 sec
Rust: 1.70 sec

Cool!

I have also removed the necessary of nightly build so every one can test and removed the-Zlocation-detail=none flag which is only supported by nightly and does nothing for performance for this type of programme and it's generally nano benefits.