@Paulo-21 2024-11-26
The second optimisation give me 
Before : 
C : 0,36 sec
Rust : 0,47 sec
After : 
Rust : 0,36 sec (same as C)