@bddicken 2024-11-29
Thanks! I was trying to figure out how to remove the nightly configuration previously but could not figure it out (I'm not a rust guy). This is great.